+++ HISTORY.md
@@ -5,7 +5,7 @@ https://github.com/josdejong/jsoneditor
 
 ## not yet released, version 5.15.1
 
-- Fixed index numbers of Array itesm not being updated after sorting.
+- Fixed index numbers of Array items not being updated after sorting.
 
 
 ## 2018-05-02, version 5.15.0